+++ typescript/cli/src/tests/warp/warp-bridge-2.e2e-test.ts
@@ -7,23 +7,10 @@ import { ChainAddresses } from '@hyperlane-xyz/registry';
 import { WarpRouteDeployConfig } from '@hyperlane-xyz/sdk';
 import { Address } from '@hyperlane-xyz/utils';
 
-import { writeYamlOrJson } from '../../utils/files.js';
-import {
-  CHAIN_NAME_2,
-  CHAIN_NAME_3,
-  DEFAULT_E2E_TEST_TIMEOUT,
-  WARP_DEPLOY_OUTPUT_PATH,
-  getCombinedWarpRoutePath,
-  sendWarpRouteMessageRoundTrip,
-} from '../commands/helpers.js';
-import { hyperlaneWarpDeploy } from '../commands/warp.js';
+import { DEFAULT_E2E_TEST_TIMEOUT } from '../commands/helpers.js';
 
-import {
-  collateralizeWarpTokens,
-  generateTestCases,
-  getTokenSymbolFromDeployment,
-  setupChains,
-} from './warp-bridge-utils.js';
+import { runWarpBridgeTests } from './warp-bridge-test-utils.js';
+import { generateTestCases, setupChains } from './warp-bridge-utils.js';
 
 chai.use(chaiAsPromised);
 chai.should();
@@ -78,62 +65,16 @@ describe('hyperlane warp deploy and bridge e2e tests - Part 2', async function (
 
   it('Should deploy and bridge different types of warp routes - Part 2:', async function () {
     this.timeout(warpConfigTestCases.length * DEFAULT_E2E_TEST_TIMEOUT);
-
-    for (let i = 0; i < warpConfigTestCases.length; i++) {
-      const warpConfig = warpConfigTestCases[i];
-      console.log(
-        `[${i + 1} of ${
-          warpConfigTestCases.length
-        }] Should deploy and be able to bridge in a ${
-          warpConfig[CHAIN_NAME_2].type
-        } -> ${warpConfig[CHAIN_NAME_3].type} warp route ...`,
-      );
-
-      writeYamlOrJson(WARP_DEPLOY_OUTPUT_PATH, warpConfig);
-      await hyperlaneWarpDeploy(WARP_DEPLOY_OUTPUT_PATH);
-
-      let startChain, targetChain: string;
-      if (!warpConfig[CHAIN_NAME_2].type.match(/.*synthetic.*/i)) {
-        startChain = CHAIN_NAME_2;
-        targetChain = CHAIN_NAME_3;
-      } else {
-        startChain = CHAIN_NAME_3;
-        targetChain = CHAIN_NAME_2;
-      }
-
-      const symbol = getTokenSymbolFromDeployment(
-        warpConfig,
-        tokenVaultChain2Symbol,
-        tokenChain2Symbol,
-        tokenVaultChain3Symbol,
-        tokenChain3Symbol,
-      );
-
-      const routeConfigPath = getCombinedWarpRoutePath(symbol, [
-        CHAIN_NAME_2,
-        CHAIN_NAME_3,
-      ]);
-
-      await collateralizeWarpTokens(routeConfigPath, warpConfig, {
-        [CHAIN_NAME_2]: {
-          wallet: walletChain2,
-          collateral: tokenChain2,
-        },
-        [CHAIN_NAME_3]: {
-          wallet: walletChain3,
-          collateral: tokenChain3,
-        },
-      });
-
-      await sendWarpRouteMessageRoundTrip(
-        startChain,
-        targetChain,
-        routeConfigPath,
-      );
-
-      console.log(
-        `Should deploy and be able to bridge in a ${warpConfig[CHAIN_NAME_2].type} -> ${warpConfig[CHAIN_NAME_3].type} warp route ✅`,
-      );
-    }
+    await runWarpBridgeTests(
+      warpConfigTestCases,
+      tokenVaultChain2Symbol,
+      tokenChain2Symbol,
+      tokenVaultChain3Symbol,
+      tokenChain3Symbol,
+      walletChain2,
+      walletChain3,
+      tokenChain2,
+      tokenChain3,
+    );
   });
 });
